Add exit when using arcanist
The `arc` command does a bunch of magic and uses git under the hood.
Because `arc` is making commits on our behalf, we want to always skip
the pre-commit hook assuming we evaluated it along the way.

This prevents us from causing failures when whole files need to be
autocorrected, but "now" is not the right time.